+++ src/RcsbSearch/SearchRequest.ts
@@ -1,11 +1,11 @@
 import {SearchQuery} from "./Types/SearchQueryInterface";
 import {QueryResult} from "./Types/SearchResultInterface";
-import * as serverSearch from "./ServerConfig/server.search.json";
+import * as serverSearch from "./ServerConfig/codegen.search.json";
 
 export class SearchRequest {
     private readonly uri: string;
     constructor(uri?:string) {
-        this.uri = uri ?? (<any>serverSearch).uri;
+        this.uri = uri ?? serverSearch.uri;
     }
     public async request(query: SearchQuery): Promise<QueryResult>{
         const response: Response = await fetch(this.uri, {