+++ src/custom.d.ts
@@ -4,6 +4,15 @@ declare module '*.graphql' {
     export = value;
 }
 declare module "*.json" {
-    const value: {schema:string;documents:string;};
+    const value: {
+        uri:string;
+        schema:string;
+        documents:string;
+        interfaceOutputFile: string;
+        pluginConfig: {
+            pluginFile:string;
+            outputFile:string;
+        }[];
+    };
     export default value;
 }